Larastan CLEAN

+++ app/Http/Controllers/ProfileController.php
@@ -26,9 +26,14 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
+        $user = $request->user();
 
-        $request->user()->save();
+        if ($user) {
+            $user->fill($request->validated());
+            $user->save();
+        } else {
+            return Redirect::route('profile.edit')->with('error', 'User not found');
+        }
 
         return Redirect::route('profile.edit')->with('status', 'profile-updated');
     }
@@ -44,9 +49,12 @@ class ProfileController extends Controller
 
         $user = $request->user();
 
-        Auth::logout();
-
-        $user->delete();
+        if ($user) {
+            Auth::logout();
+            $user->delete();
+        } else {
+            return Redirect::route('profile.edit')->with('error', 'User not found');
+        }
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
